Move mail from services to modules
